IPA: Make ipa_hbac_sysdb_save() more generic
Although there's no change in the ipa_hbac_sysdb_save() itself, its name
has been changed to ipa_common_entries_and_groups_sysdb_save() and its
been split out from HBAC related files and moved to the newly created
ipa_rules_common.[ch] files, which will also be used in the future for
new backend modules.

ipa_rules_common.[ch] is not exactly the best name for those files, IMO,
but I really cannot come up with something better.

Related:
https://pagure.io/SSSD/sssd/issue/2995
